Make SDC RefEnumerable, upgrade project to .NET9
The ref enumerable was totally required to create totally necessary performance improvements.
Actually it was quite necessary for allowing SDC's to be used in a more data oriented way - read only memory is quite useless to have publicly available in this way.

This did reveal a little fail in the style guide. So that's something